+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/organisation/OrganisationServiceClient.java
@@ -1,5 +1,7 @@
 package dk.medcom.video.api.organisation;
 
+import dk.medcom.video.api.organisation.model.Organisation;
+
 import java.util.List;
 
 public interface OrganisationServiceClient {
