+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/organisation/OrganisationServiceStrategy.java
@@ -1,5 +1,7 @@
 package dk.medcom.video.api.organisation;
 
+import dk.medcom.video.api.organisation.model.Organisation;
+
 import javax.ws.rs.NotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
