+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/organisation/OrganisationTreeServiceClient.java
@@ -1,5 +1,7 @@
 package dk.medcom.video.api.organisation;
 
+import dk.medcom.video.api.organisation.model.OrganisationTree;
+
 public interface OrganisationTreeServiceClient {
     OrganisationTree getOrganisationTree(String organisationCode);
 }
