+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/service/impl/OrganisationFinder.java
@@ -1,6 +1,6 @@
 package dk.medcom.video.api.service.impl;
 
-import dk.medcom.video.api.organisation.OrganisationTree;
+import dk.medcom.video.api.organisation.model.OrganisationTree;
 
 import java.util.Optional;
