+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/interceptor/OrganisationInterceptorTest.java
@@ -4,7 +4,7 @@ import dk.medcom.video.api.context.UserContext;
 import dk.medcom.video.api.context.UserContextImpl;
 import dk.medcom.video.api.context.UserContextService;
 import dk.medcom.video.api.context.UserRole;
-import dk.medcom.video.api.organisation.Organisation;
+import dk.medcom.video.api.organisation.model.Organisation;
 import dk.medcom.video.api.organisation.OrganisationServiceClient;
 import dk.medcom.video.api.organisation.OrganisationStrategy;
 import dk.medcom.video.api.dao.OrganisationRepository;
