+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/service/impl/OrganisationFinderTest.java
@@ -1,7 +1,6 @@
 package dk.medcom.video.api.service.impl;
 
-import dk.medcom.video.api.organisation.OrganisationTree;
-import dk.medcom.video.api.service.impl.OrganisationFinder;
+import dk.medcom.video.api.organisation.model.OrganisationTree;
 import org.junit.Before;
 import org.junit.Test;
