+++ medcom-video-api-test/src/main/java/dk/medcom/video/api/TestApplication.java
@@ -5,8 +5,8 @@ import com.github.dockerjava.api.model.ExposedPort;
 import com.github.dockerjava.api.model.PortBinding;
 import com.github.dockerjava.api.model.Ports;
 import com.google.common.net.MediaType;
-import dk.medcom.video.api.organisation.Organisation;
-import dk.medcom.video.api.organisation.OrganisationTree;
+import dk.medcom.video.api.organisation.model.Organisation;
+import dk.medcom.video.api.organisation.model.OrganisationTree;
 import org.mockserver.client.server.MockServerClient;
 import org.mockserver.model.Header;
 import org.mockserver.model.HttpRequest;
